Made login and logout part of Makefile
